requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
beautifulsoup4==4.12.2
aiohttp==3.8.4
python-dotenv==1.0.0
pydantic==1.10.7
tenacity==8.2.2
pytest==7.4.0

langchain==0.0.307
openai==0.27.8
